chore: bunch more dead code, god, cursor is the worst
